windsurf-safari-assistant/windsurf-safari-01/CascadeProjects/windsurf-project/requirements.txt
markdown==3.5.1
python-frontmatter==1.0.0
click==8.1.7
rich==13.7.0
